Install height provider next to the sheet's content
